Add SynchronousBuild
https://github.com/OpenRA/OpenRA/pull/14096

+++ OpenRA.Mods.Common/Widgets/ObserverProductionIconsWidget.cs
@@ -141,7 +141,7 @@ namespace OpenRA.Mods.Common.Widgets
 
 				var clock = clocks[queue.Trait];
 				clock.PlayFetchIndex(ClockSequence,
-					() => current.TotalTime == 0 ? 0 : ((current.TotalTime - current.RemainingTime)
+					() => current.TotalTime == 0 ? 0 : ((current.TotalTime - (int)current.RemainingTime)
 					* (clock.CurrentSequence.Length - 1) / current.TotalTime));
 				clock.Tick();
 				WidgetUtils.DrawSHPCentered(clock.Image, location + 0.5f * iconSize, worldRenderer.Palette(ClockPalette), 0.5f);
